Fixes crashes due to use of the deprecated TaskInternal.execute() method and provides support for multiple sets of includes/excludes
