+++ src/main/groovy/com/radcortez/gradle/plugin/openjpa/OpenJpaPlugin.groovy
@@ -1,14 +1,12 @@
 package com.radcortez.gradle.plugin.openjpa
 
 import com.radcortez.gradle.plugin.openjpa.enhance.EnhanceExtension
-import com.radcortez.gradle.plugin.openjpa.enhance.EnhanceTask
 import com.radcortez.gradle.plugin.openjpa.metamodel.MetamodelExtension
 import com.radcortez.gradle.plugin.openjpa.metamodel.MetamodelTask
 import com.radcortez.gradle.plugin.openjpa.sql.SqlExtension
 import com.radcortez.gradle.plugin.openjpa.sql.SqlTask
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-
 /**
  * Description.
  *
@@ -21,13 +19,6 @@ class OpenJpaPlugin implements Plugin<Project> {
         project.openjpa.extensions.create("enhance", EnhanceExtension, project)
         project.openjpa.extensions.create("metamodel", MetamodelExtension, project)
         project.openjpa.extensions.create("sql", SqlExtension, project)
-
-        project.task(
-                type: EnhanceTask,
-                group: "OpenJPA",
-                description: "Enhances entity classes with the OpenJPA Enhancer tool.",
-                dependsOn: "classes",
-                "enhance")
 
         project.task(
                 type: MetamodelTask,
